tikv_util: move SegmentSet into tikv_util
