+++ components/tikv_util/src/lib.rs
@@ -53,6 +53,7 @@ pub mod memory;
 pub mod metrics;
 pub mod mpsc;
 pub mod quota_limiter;
+pub mod range;
 pub mod resource_control;
 pub mod store;
 pub mod stream;
